models/product_configuration.py
```python
from odoo import models, fields, api

class ProductConfiguration(models.Model):
    _name = 'product.configuration'
    _description = 'Product Configuration'

    name = fields.Char(string='Name', required=True)
    cpu = fields.Many2one('product.template', string='CPU')
    gpu = fields.Many2one('product.template', string='GPU')
    ram = fields.Many2one('product.template', string='RAM')
    storage = fields.Many2one('product.template', string='Storage')
    power_supply = fields.Many2one('product.template', string='Power Supply')
    case = fields.Many2one('product.template', string='Case')

    @api.onchange('cpu', 'gpu', 'ram', 'storage', 'power_supply', 'case')
    def _check_compatibility(self):
        # Implement compatibility logic here
        pass

    @api.model
    def create(self, vals):
        # Implement additional logic here
        record = super().create(vals)
        return record

    def write(self, vals):
        # Implement additional logic here
        result = super().write(vals)
        return result

    def unlink(self):
        # Implement additional logic here
        result = super().unlink()
        return result
```